Create base test for testing all the generic editor screen.
